+++ src/ExpenseTrackerGrupo2/DataAccess/Concretes/Interfaces/IBaseRepository.cs
@@ -1,4 +1,4 @@
-namespace ExpenseTrackerGrupo2.src.ExpenseTrackerGrupo2.DataAccess.Layer;
+namespace ExpenseTrackerGrupo2.src.ExpenseTrackerGrupo2.DataAccess.Concretes;
 
 public interface IBaseRepository<T>
 {
