+++ src/ExpenseTrackerGrupo2/DataAccess/Concretes/Interfaces/IExpenseRepository.cs
@@ -1,6 +1,6 @@
 using ExpenseTrackerGrupo2.src.ExpenseTrackerGrupo2.DataAccess.Entities;
 
-namespace ExpenseTrackerGrupo2.src.ExpenseTrackerGrupo2.DataAccess.Layer;
+namespace ExpenseTrackerGrupo2.src.ExpenseTrackerGrupo2.DataAccess.Concretes;
 public interface IExpenseRepository : IBaseRepository<Expense>
 {
     Task<IList<Expense>> GetExpenseByCategory(string category);
